Rebases against main and updates tests accordingly
