+++ server/utils/assessmentUtils.test.ts
@@ -3,6 +3,7 @@ import {
   assessmentDataNoAssessments,
   assessmentUndefined,
   crimNeeds,
+  crimNeedsOrdering,
 } from '../testutils/data/assessmentData'
 import locale from '../routes/aboutPop/locale.json'
 import {
@@ -72,55 +73,55 @@ describe('format assessment data', () => {
         other: [
           {
             goalRoute: 'alcohol-use',
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-            riskOfReoffending: undefined,
             thresholdValue: 1,
             title: 'Alcohol use',
             upperBound: 4,
+            linkedtoReoffending: undefined,
+            linkedtoRoSH: undefined,
+            riskOfReoffending: undefined,
           },
           {
             goalRoute: 'drug-use',
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-            riskOfReoffending: undefined,
             thresholdValue: 0,
             title: 'Drug use',
             upperBound: 8,
+            linkedtoReoffending: undefined,
+            linkedtoRoSH: undefined,
+            riskOfReoffending: undefined,
           },
           {
             goalRoute: 'finances',
+            title: 'Finances',
+            upperBound: null,
             linkedtoReoffending: undefined,
             linkedtoRoSH: undefined,
             riskOfReoffending: undefined,
-            title: 'Finances',
-            upperBound: null,
           },
           {
             goalRoute: 'health-and-wellbeing',
+            title: 'Health and wellbeing',
+            upperBound: null,
             linkedtoReoffending: false,
             linkedtoRoSH: false,
             riskOfReoffending: undefined,
-            title: 'Health and wellbeing',
-            upperBound: null,
           },
           {
             goalRoute: 'personal-relationships-and-community',
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-            riskOfReoffending: undefined,
             thresholdValue: 1,
             title: 'Personal relationships and community',
             upperBound: 6,
-          },
-          {
-            goalRoute: 'thinking-behaviours-and-attitudes',
             linkedtoReoffending: undefined,
             linkedtoRoSH: undefined,
             riskOfReoffending: undefined,
+          },
+          {
+            goalRoute: 'thinking-behaviours-and-attitudes',
             thresholdValue: 2,
             title: 'Thinking, behaviours and attitudes',
             upperBound: 10,
+            linkedtoReoffending: undefined,
+            linkedtoRoSH: undefined,
+            riskOfReoffending: undefined,
           },
         ],
         versionUpdatedAt: '2024-10-04T15:22:31.453096',
@@ -162,55 +163,144 @@ describe('format assessment data', () => {
         other: [
           {
             goalRoute: 'alcohol-use',
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-            riskOfReoffending: undefined,
             thresholdValue: 1,
             title: 'Alcohol use',
             upperBound: 4,
-          },
-          {
-            goalRoute: 'drug-use',
             linkedtoReoffending: undefined,
             linkedtoRoSH: undefined,
             riskOfReoffending: undefined,
+          },
+          {
+            goalRoute: 'drug-use',
+            thresholdValue: 0,
+            title: 'Drug use',
+            upperBound: 8,
+            linkedtoReoffending: undefined,
+            linkedtoRoSH: undefined,
+            riskOfReoffending: undefined,
+          },
+          {
+            goalRoute: 'finances',
+            title: 'Finances',
+            upperBound: null,
+            linkedtoReoffending: undefined,
+            linkedtoRoSH: undefined,
+            riskOfReoffending: undefined,
+          },
+          {
+            goalRoute: 'health-and-wellbeing',
+            title: 'Health and wellbeing',
+            upperBound: null,
+            linkedtoReoffending: false,
+            linkedtoRoSH: false,
+            riskOfReoffending: undefined,
+          },
+          {
+            goalRoute: 'personal-relationships-and-community',
+            thresholdValue: 1,
+            title: 'Personal relationships and community',
+            upperBound: 6,
+            linkedtoReoffending: undefined,
+            linkedtoRoSH: undefined,
+            riskOfReoffending: undefined,
+          },
+          {
+            goalRoute: 'thinking-behaviours-and-attitudes',
+            thresholdValue: 2,
+            title: 'Thinking, behaviours and attitudes',
+            upperBound: 10,
+            linkedtoReoffending: undefined,
+            linkedtoRoSH: undefined,
+            riskOfReoffending: undefined,
+          },
+        ],
+        versionUpdatedAt: '2024-10-04T15:22:31.453096',
+      },
+    ],
+    [
+      crimNeedsOrdering,
+      assessmentDataNoAssessments,
+      locale.en.areas,
+      {
+        highScoring: [
+          {
+            criminogenicNeedsScore: '4',
+            goalRoute: 'drug-use',
+            linkedtoReoffending: false,
+            linkedtoRoSH: false,
+            riskOfReoffending: undefined,
+            overallScore: '4',
             thresholdValue: 0,
             title: 'Drug use',
             upperBound: 8,
           },
           {
-            goalRoute: 'finances',
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-            riskOfReoffending: undefined,
-            title: 'Finances',
-            upperBound: null,
-          },
-          {
-            goalRoute: 'health-and-wellbeing',
+            criminogenicNeedsScore: '5',
+            goalRoute: 'thinking-behaviours-and-attitudes',
             linkedtoReoffending: false,
             linkedtoRoSH: false,
             riskOfReoffending: undefined,
-            title: 'Health and wellbeing',
-            upperBound: null,
-          },
-          {
-            goalRoute: 'personal-relationships-and-community',
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-            riskOfReoffending: undefined,
-            thresholdValue: 1,
-            title: 'Personal relationships and community',
-            upperBound: 6,
-          },
-          {
-            goalRoute: 'thinking-behaviours-and-attitudes',
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-            riskOfReoffending: undefined,
+            overallScore: '5',
             thresholdValue: 2,
             title: 'Thinking, behaviours and attitudes',
             upperBound: 10,
+          },
+          {
+            criminogenicNeedsScore: '3',
+            goalRoute: 'accommodation',
+            linkedtoReoffending: false,
+            linkedtoRoSH: false,
+            riskOfReoffending: undefined,
+            overallScore: '3',
+            thresholdValue: 1,
+            title: 'Accommodation',
+            upperBound: 6,
+          },
+        ],
+        lowScoring: [],
+        other: [
+          {
+            goalRoute: 'alcohol-use',
+            thresholdValue: 1,
+            title: 'Alcohol use',
+            upperBound: 4,
+            linkedtoRoSH: undefined,
+            riskOfReoffending: undefined,
+            linkedtoReoffending: undefined,
+          },
+          {
+            goalRoute: 'employment-and-education',
+            thresholdValue: 1,
+            title: 'Employment and education',
+            linkedtoRoSH: undefined,
+            riskOfReoffending: undefined,
+            linkedtoReoffending: undefined,
+            upperBound: 4,
+          },
+          {
+            goalRoute: 'finances',
+            title: 'Finances',
+            upperBound: null,
+            linkedtoRoSH: undefined,
+            riskOfReoffending: undefined,
+            linkedtoReoffending: undefined,
+          },
+          {
+            goalRoute: 'health-and-wellbeing',
+            title: 'Health and wellbeing',
+            upperBound: null,
+            linkedtoRoSH: undefined,
+            riskOfReoffending: undefined,
+            linkedtoReoffending: undefined,
+          },
+          {
+            goalRoute: 'personal-relationships-and-community',
+            thresholdValue: 1,
+            title: 'Personal relationships and community',
+            upperBound: 6,
+            linkedtoRoSH: undefined,
+            riskOfReoffending: undefined,
+            linkedtoReoffending: undefined,
           },
         ],
         versionUpdatedAt: '2024-10-04T15:22:31.453096',
