+++ server/utils/utils.test.ts
@@ -5,28 +5,16 @@ import {
   convertToTitleCase,
   dateToISOFormat,
   dateWithYear,
-  formatAssessmentData,
   formatDate,
   formatRoSHData,
   goalStatusToTabName,
   initialiseName,
-  motivationText,
   moveGoal,
   sortSteps,
   toKebabCase,
-  yearsAndDaysElapsed,
 } from './utils'
 import { NewStep, StepStatus } from '../@types/StepType'
 import { GoalStatus } from '../@types/GoalType'
-import { AssessmentAreaConfig, AssessmentAreas, AssessmentResponse, CriminogenicNeedsData } from '../@types/Assessment'
-import {
-  assessmentData,
-  assessmentDataNoAssessments,
-  assessmentUndefined,
-  crimNeeds,
-  crimNeedsOrdering,
-} from '../testutils/data/assessmentData'
-import locale from '../routes/aboutPop/locale.json'
 
 describe('convert to title case', () => {
   it.each([
@@ -204,327 +192,6 @@ describe('format dates correctly', () => {
     expect(dateToISOFormat('31/3/2023')).toEqual('2023-03-31')
     expect(dateToISOFormat('1/3/2023')).toEqual('2023-03-01')
     expect(dateToISOFormat('  1/3/2023')).toEqual('2023-03-01')
-  })
-})
-
-describe('format assessment data', () => {
-  it.each([
-    [
-      crimNeeds,
-      assessmentUndefined,
-      locale.en.areas,
-      {
-        highScoring: [],
-        lowScoring: [],
-        other: [],
-        versionUpdatedAt: undefined,
-      },
-    ],
-    [
-      crimNeeds,
-      null,
-      locale.en.areas,
-      {
-        highScoring: [],
-        lowScoring: [],
-        other: [],
-        versionUpdatedAt: undefined,
-      },
-    ],
-    [
-      crimNeeds,
-      assessmentDataNoAssessments,
-      locale.en.areas,
-      {
-        highScoring: [
-          {
-            criminogenicNeedsScore: '6',
-            linkedtoReoffending: true,
-            linkedtoRoSH: false,
-            overallScore: '6',
-            riskOfReoffending: undefined,
-            thresholdValue: 1,
-            title: 'Accommodation',
-            goalRoute: 'accommodation',
-            upperBound: 6,
-          },
-        ],
-        lowScoring: [
-          {
-            criminogenicNeedsScore: '1',
-            linkedtoReoffending: true,
-            linkedtoRoSH: false,
-            overallScore: '1',
-            riskOfReoffending: undefined,
-            thresholdValue: 1,
-            title: 'Employment and education',
-            goalRoute: 'employment-and-education',
-            upperBound: 4,
-          },
-        ],
-        other: [
-          {
-            goalRoute: 'thinking-behaviours-and-attitudes',
-            thresholdValue: 2,
-            title: 'Thinking, behaviours and attitudes',
-            upperBound: 10,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-            riskOfReoffending: undefined,
-          },
-          {
-            goalRoute: 'alcohol-use',
-            thresholdValue: 1,
-            title: 'Alcohol use',
-            upperBound: 4,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-            riskOfReoffending: undefined,
-          },
-          {
-            goalRoute: 'health-and-wellbeing',
-            linkedtoReoffending: false,
-            linkedtoRoSH: false,
-            riskOfReoffending: undefined,
-            title: 'Health and wellbeing',
-            upperBound: null,
-          },
-          {
-            goalRoute: 'finances',
-            title: 'Finances',
-            upperBound: null,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-            riskOfReoffending: undefined,
-          },
-          {
-            goalRoute: 'personal-relationships-and-community',
-            thresholdValue: 1,
-            title: 'Personal relationships and community',
-            upperBound: 6,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-            riskOfReoffending: undefined,
-          },
-          {
-            goalRoute: 'drug-use',
-            thresholdValue: 0,
-            title: 'Drug use',
-            upperBound: 8,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-            riskOfReoffending: undefined,
-          },
-        ],
-        versionUpdatedAt: '2024-10-04T15:22:31.453096',
-      },
-    ],
-    [
-      crimNeeds,
-      assessmentData,
-      locale.en.areas,
-      {
-        highScoring: [
-          {
-            criminogenicNeedsScore: '6',
-            linkedtoReoffending: true,
-            linkedtoRoSH: false,
-            motivationToMakeChanges: 'thinkingAboutMakingChanges',
-            overallScore: '6',
-            riskOfReoffending: undefined,
-            goalRoute: 'accommodation',
-            thresholdValue: 1,
-            title: 'Accommodation',
-            upperBound: 6,
-          },
-        ],
-        lowScoring: [
-          {
-            criminogenicNeedsScore: '1',
-            linkedtoReoffending: true,
-            linkedtoRoSH: false,
-            motivationToMakeChanges: 'needsHelpToMakeChanges',
-            overallScore: '1',
-            riskOfReoffending: undefined,
-            thresholdValue: 1,
-            title: 'Employment and education',
-            goalRoute: 'employment-and-education',
-            upperBound: 4,
-          },
-        ],
-        other: [
-          {
-            goalRoute: 'thinking-behaviours-and-attitudes',
-            thresholdValue: 2,
-            title: 'Thinking, behaviours and attitudes',
-            upperBound: 10,
-            riskOfReoffending: undefined,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-          },
-          {
-            goalRoute: 'alcohol-use',
-            thresholdValue: 1,
-            title: 'Alcohol use',
-            upperBound: 4,
-            riskOfReoffending: undefined,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-          },
-          {
-            goalRoute: 'health-and-wellbeing',
-            linkedtoReoffending: false,
-            linkedtoRoSH: false,
-            title: 'Health and wellbeing',
-            upperBound: null,
-            riskOfReoffending: undefined,
-          },
-          {
-            goalRoute: 'finances',
-            title: 'Finances',
-            upperBound: null,
-            riskOfReoffending: undefined,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-          },
-          {
-            goalRoute: 'personal-relationships-and-community',
-            thresholdValue: 1,
-            title: 'Personal relationships and community',
-            upperBound: 6,
-            riskOfReoffending: undefined,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-          },
-          {
-            goalRoute: 'drug-use',
-            thresholdValue: 0,
-            title: 'Drug use',
-            upperBound: 8,
-            riskOfReoffending: undefined,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-          },
-        ],
-        versionUpdatedAt: '2024-10-04T15:22:31.453096',
-      },
-    ],
-    [
-      crimNeedsOrdering,
-      assessmentDataNoAssessments,
-      locale.en.areas,
-      {
-        highScoring: [
-          {
-            goalRoute: 'drug-use',
-            criminogenicNeedsScore: '4',
-            overallScore: '4',
-            linkedtoReoffending: false,
-            linkedtoRoSH: false,
-            riskOfReoffending: undefined,
-            thresholdValue: 0,
-            title: 'Drug use',
-            upperBound: 8,
-          },
-          {
-            goalRoute: 'thinking-behaviours-and-attitudes',
-            criminogenicNeedsScore: '5',
-            overallScore: '5',
-            linkedtoReoffending: false,
-            linkedtoRoSH: false,
-            riskOfReoffending: undefined,
-            thresholdValue: 2,
-            title: 'Thinking, behaviours and attitudes',
-            upperBound: 10,
-          },
-          {
-            goalRoute: 'accommodation',
-            criminogenicNeedsScore: '3',
-            overallScore: '3',
-            linkedtoReoffending: false,
-            linkedtoRoSH: false,
-            riskOfReoffending: undefined,
-            thresholdValue: 1,
-            title: 'Accommodation',
-            upperBound: 6,
-          },
-        ],
-        lowScoring: [],
-        other: [
-          {
-            goalRoute: 'alcohol-use',
-            thresholdValue: 1,
-            title: 'Alcohol use',
-            upperBound: 4,
-            riskOfReoffending: undefined,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-          },
-          {
-            goalRoute: 'health-and-wellbeing',
-            title: 'Health and wellbeing',
-            upperBound: null,
-            riskOfReoffending: undefined,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-          },
-          {
-            goalRoute: 'employment-and-education',
-            thresholdValue: 1,
-            title: 'Employment and education',
-            upperBound: 4,
-            riskOfReoffending: undefined,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-          },
-          {
-            goalRoute: 'finances',
-            title: 'Finances',
-            upperBound: null,
-            riskOfReoffending: undefined,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-          },
-          {
-            goalRoute: 'personal-relationships-and-community',
-            thresholdValue: 1,
-            title: 'Personal relationships and community',
-            upperBound: 6,
-            riskOfReoffending: undefined,
-            linkedtoReoffending: undefined,
-            linkedtoRoSH: undefined,
-          },
-        ],
-        versionUpdatedAt: '2024-10-04T15:22:31.453096',
-      },
-    ],
-  ])(
-    '%s formatAssessmentData(%s, %s, %s, %s',
-    (
-      criminogenicNeedsData: CriminogenicNeedsData,
-      assessment: AssessmentResponse,
-      areas: AssessmentAreaConfig[],
-      expected: AssessmentAreas,
-    ) => {
-      expect(formatAssessmentData(criminogenicNeedsData, assessment, areas)).toEqual(expected)
-    },
-  )
-})
-
-describe('replace motivation text', () => {
-  it.each([
-    ['MADE_CHANGES', 'madeChanges'],
-    ['MAKING_CHANGES', 'makingChanges'],
-    ['WANT_TO_MAKE_CHANGES', 'wantToMakeChanges'],
-    ['NEEDS_HELP_TO_MAKE_CHANGES', 'needsHelpToMakeChanges'],
-    ['THINKING_ABOUT_MAKING_CHANGES', 'thinkingAboutMakingChanges'],
-    ['DOES_NOT_WANT_TO_MAKE_CHANGES', 'doesNotWantToMakeChanges'],
-    ['DOES_NOT_WANT_TO_ANSWER', 'doesNotWantToAnswer'],
-    ['NOT_PRESENT', 'notPresent'],
-    ['NOT_APPLICABLE', 'notApplicable'],
-    [undefined, undefined],
-  ])('%s %s maps to %s', (optionResult: string, expected: string) => {
-    expect(motivationText(optionResult)).toEqual(expected)
   })
 })
